+++ include/Small/All.hpp
@@ -16,5 +16,6 @@
 #include <Small/Graphics.hpp>
 #include <Small/System.hpp>
 #include <Small/Scenes.hpp>
+#include <Small/Widgets.hpp>
 
 #endif