+++ include/Small/Widgets.hpp
@@ -1,0 +1,15 @@
+/**
+ * @file Widgets.hpp
+ * @author Alexandre Plateau (lexplt.dev@gmail.com)
+ * @brief Include all the files from the Widgets folder
+ * @version 0.1
+ * @date 2020-04-15
+ * 
+ * @copyright Copyright (c) 2020
+ * 
+ */
+
+#ifndef sgl_small_widgets
+#define sgl_small_widgets
+
+#endif