+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - added an example of how to use `sgl::Graphics::TileView`
 - adding a sgl::Scenes::Zone to handle zone transitions on a map
 - added `Small/System.hpp` and `Small/Scenes.hpp` to `Small/All.hpp`
+- adding a base widget to be completed
 
 ### Changed
 - the tilemap has been updated to no longer hold the texture, and be able to modify the vertex array when needed